Update to use training set and test set
splits data3 up int approx 75% of data to train and the remainder to test.

+++ Solve_data3/src/bioassignment/Data.java
@@ -22,7 +22,7 @@ public class Data {
     @Override
     public  String toString(){
         String s = "";
-        for(float v:variables){
+        for(double v:variables){
             s = s+v;
         }
       return  s+type;
